# Improve performance of product category rewrites (disabled product & categories and product_use_categories)