--- a/app/code/core/Mage/Catalog/Model/Indexer/Url.php
+++ b/app/code/core/Mage/Catalog/Model/Indexer/Url.php
@@ -47,7 +47,8 @@ class Mage_Catalog_Model_Indexer_Url extends Mage_Index_Model_Indexer_Abstract
      */
     protected $_matchedEntities = [
         Mage_Catalog_Model_Product::ENTITY => [
-            Mage_Index_Model_Event::TYPE_SAVE
+            Mage_Index_Model_Event::TYPE_SAVE,
+            Mage_Index_Model_Event::TYPE_MASS_ACTION
         ],
         Mage_Catalog_Model_Category::ENTITY => [
             Mage_Index_Model_Event::TYPE_SAVE
@@ -180,13 +181,36 @@ class Mage_Catalog_Model_Indexer_Url extends Mage_Index_Model_Indexer_Abstract
     protected function _registerProductEvent(Mage_Index_Model_Event $event)
     {
         $product = $event->getDataObject();
-        $dataChange = $product->dataHasChangedFor('url_key')
-            || $product->dataHasChangedFor('status')
-            || $product->getIsChangedCategories()
-            || $product->getIsChangedWebsites();
+        $dataChange = false;
+        $products = [];
 
-        if (!$product->getExcludeUrlRewrite() && $dataChange) {
-            $event->addNewData('rewrite_product_ids', [$product->getId()]);
+        if (is_a($product, Mage_Catalog_Model_Product_Action::class)) {
+            $attributesData = $product->getData('attributes_data');
+            $productsIds = $product->getData('product_ids');
+            $dataChange = isset($attributesData['status']) && isset($productsIds) && count($productsIds) > 0;
+            if ($dataChange) {
+                $products = Mage::getModel('catalog/product')->getCollection()
+                    ->addFieldToFilter('entity_id', array('in'=> $productsIds));
+            } else {
+                return;
+            }
+        }
+
+        if (is_a($product, Mage_Catalog_Model_Product::class)) {
+            $dataChange =
+                (
+                    $product->dataHasChangedFor('url_key')
+                    || $product->dataHasChangedFor('status')
+                    || $product->getIsChangedCategories()
+                    || $product->getIsChangedWebsites()
+                ) && !$product->getExcludeUrlRewrite();
+            $products = [$product];
+        }
+
+        if ($dataChange) {
+            foreach ($products as $product) {
+                $event->addNewData('rewrite_product_ids', [$product->getId()]);
+            }
         }
     }