--- a/app/code/core/Mage/Catalog/Model/Indexer/Url.php
+++ b/app/code/core/Mage/Catalog/Model/Indexer/Url.php
@@ -181,6 +181,7 @@ class Mage_Catalog_Model_Indexer_Url extends Mage_Index_Model_Indexer_Abstract
     {
         $product = $event->getDataObject();
         $dataChange = $product->dataHasChangedFor('url_key')
+            || $product->dataHasChangedFor('status')
             || $product->getIsChangedCategories()
             || $product->getIsChangedWebsites();
 
@@ -198,7 +199,11 @@ class Mage_Catalog_Model_Indexer_Url extends Mage_Index_Model_Indexer_Abstract
     {
         $category = $event->getDataObject();
         if (!$category->getInitialSetupFlag() && $category->getLevel() > 1) {
-            if ($category->dataHasChangedFor('url_key') || $category->getIsChangedProductList()) {
+            if (
+                $category->dataHasChangedFor('url_key')
+                || $category->getIsChangedProductList()
+                || $category->dataHasChangedFor('is_active')
+            ) {
                 $event->addNewData('rewrite_category_ids', [$category->getId()]);
             }
             /**