--- a/app/code/core/Mage/Catalog/Model/Url.php
+++ b/app/code/core/Mage/Catalog/Model/Url.php
@@ -112,11 +112,14 @@ class Mage_Catalog_Model_Url
     protected $productUseCategories;
 
     /**
-     * Neuca_Base_Model_Url constructor.
+     * @var bool
      */
+    protected $createForDisabled;
+
     public function __construct()
     {
         $this->productUseCategories = Mage::getStoreConfigFlag('catalog/seo/product_use_categories');
+        $this->createForDisabled = Mage::getStoreConfigFlag('catalog/seo/create_url_for_disabled');
     }
 
     /**
@@ -258,7 +261,7 @@ class Mage_Catalog_Model_Url
         $this->clearStoreInvalidRewrites($storeId);
         $this->refreshCategoryRewrite($this->getStores($storeId)->getRootCategoryId(), $storeId, false);
         $this->refreshProductRewrites($storeId);
-        $this->getResource()->clearCategoryProduct($storeId);
+        $this->getResource()->clearRewrites($storeId);
 
         return $this;
     }
@@ -384,7 +387,7 @@ class Mage_Catalog_Model_Url
     }
 
     /**
-     * Refresh products for catwgory
+     * Refresh products for category
      *
      * @param Varien_Object|Mage_Catalog_Model_Category $category
      * @return $this
@@ -441,7 +444,7 @@ class Mage_Catalog_Model_Url
      */
     public function refreshCategoryRewrite($categoryId, $storeId = null, $refreshProducts = null)
     {
-        if (!$refreshProducts) {
+        if (is_null($refreshProducts)) {
             $refreshProducts = $this->productUseCategories;
         }
         if (is_null($storeId)) {
@@ -453,6 +456,11 @@ class Mage_Catalog_Model_Url
 
         $category = $this->getResource()->getCategory($categoryId, $storeId);
         if (!$category) {
+            return $this;
+        }
+
+        if (!$this->createForDisabled && !$category->getIsActive()) {
+            $this->getResource()->clearDisabledCategory($category->getId());
             return $this;
         }
 
@@ -488,7 +496,7 @@ class Mage_Catalog_Model_Url
             return $this;
         }
 
-        $product = $this->getResource()->getProduct($productId, $storeId);
+        $product = $this->getResource()->getProduct($productId, $storeId, $this->createForDisabled);
         if (!$product) {
             // Product doesn't belong to this store - clear all its url rewrites including root one
             $this->getResource()->clearProductRewrites($productId, $storeId, array());
@@ -520,6 +528,11 @@ class Mage_Catalog_Model_Url
 
         // Remove all other product rewrites created earlier for this store - they're invalid now
         $excludeCategoryIds = array_keys($categories);
+
+        // Product is disabled and in configuration set to not create for disabled - clear all its url rewrites including root one
+        if (!$this->createForDisabled && $product->getStatus() === Mage_Catalog_Model_Product_Status::STATUS_DISABLED) {
+            $excludeCategoryIds = [];
+        }
         $this->getResource()->clearProductRewrites($productId, $storeId, $excludeCategoryIds);
 
         unset($categories);
@@ -545,7 +558,8 @@ class Mage_Catalog_Model_Url
         $process = true;
 
         while ($process == true) {
-            $products = $this->getResource()->getProductsByStore($storeId, $lastEntityId);
+            $products = $this->getResource()->getProductsByStore($storeId, $lastEntityId, $this->createForDisabled);
+
             if (!$products) {
                 $process = false;
                 break;
@@ -565,7 +579,7 @@ class Mage_Catalog_Model_Url
                 }
 
                 if ($loadCategories) {
-                    foreach ($this->getResource()->getCategories($loadCategories, $storeId) as $category) {
+                    foreach ($this->getResource()->getCategories($loadCategories, $storeId, $this->createForDisabled) as $category) {
                         $this->_categories[$category->getId()] = $category;
                     }
                 }