… app/code/core/Mage/Catalog/etc/config.xml
… app/code/core/Mage/Catalog/etc/system.xml