@MarcinNowakMacopedia 2022-04-27
> The only change I would recommend, thinking of a wider audience, is to display a message, when switching `product_use_categories` from `No` to `Yes`, that reindexing URL Rewrites is recommended. I was even thinking that the original intention of still generating URLs was for a quicker transition and not actually an oversight (arguable but possible).

Added