@tmotyl 2022-06-21
fixed  issues reported by phpstan