@colinmollenhour 2022-08-03
Anyone want to test this on their staging environments and report back? I currently don't have a production-size instance to test with.