@MarcinNowakMacopedia 2022-08-19
@colinmollenhour I added Mage_Adminhtml_Model_System_Config_Backend_Seo to change  the index status of categories and products after changing create_url_for_disabled in configuration. Additionally, I modified  Mage_Adminhtml_Model_System_Config_Backend_Seo_Product to change product indexer status, after changing product_use_categories.
https://github.com/OpenMage/magento-lts/pull/1405/commits/e33773c8c1e295271457c66bb14cf682b53f4d5b