@MarcinNowakMacopedia 2022-08-29
@fballiano I resolved conflicts.