@colinmollenhour 2023-01-19
@MarcinNowakMacopedia Are you able to look at my PR https://github.com/macopedia/magento-lts/pull/8? I am not able to push directly to your branch.